demo/requirements.txt
gradio_rag_sources